+++ src/Container/DefaultServiceContainer.php
@@ -2,8 +2,10 @@
 
 namespace Solital\Core\Container;
 
+use Solital\Core\Auth\Password;
 use Solital\Core\Container\Interface\{ContainerInterface, ServiceProviderInterface};
 use Solital\Core\FileSystem\HandleFiles;
+use Solital\Core\Wolf\Wolf;
 
 class DefaultServiceContainer implements ServiceProviderInterface
 {
@@ -15,9 +17,8 @@ class DefaultServiceContainer implements ServiceProviderInterface
     #[\Override]
     public function register(ContainerInterface $container)
     {
-        /* $container->add('handler-file', function() {
-            return new HandleFiles;
-        }); */
         $container->add('handler-file', fn() => new HandleFiles);
+        $container->add('solital-password', fn() => new Password);
+        $container->add('solital-wolf', fn() => new Wolf);
     }
 }