+++ src/Kernel/Console/Commands/HandleCache.php
@@ -5,7 +5,6 @@ namespace Solital\Core\Kernel\Console\Commands;
 use Solital\Core\Console\Command;
 use Solital\Core\Kernel\Application;
 use Solital\Core\Console\Interface\CommandInterface;
-use Solital\Core\FileSystem\HandleFiles;
 
 class HandleCache extends Command implements CommandInterface
 {