+++ src/Resource/Helpers/security.php
@@ -1,7 +1,7 @@
 <?php
 
-use Solital\Core\Auth\Password;
 use Solital\Core\Course\Course as Course;
+use Solital\Core\Kernel\Application;
 
 /**
  * Get current csrf-token
@@ -35,7 +35,7 @@ function spoofing(string $method): string
  */
 function pass_hash(string $value, bool $info = false): string
 {
-    return (new Password())->create($value, $info);
+    return Application::provider('solital-password')->create($value, $info);
 }
 
 /**
@@ -44,5 +44,5 @@ function pass_hash(string $value, bool $info = false): string
  */
 function pass_verify(string $value, string $hash)
 {
-    return (new Password())->verify($value, $hash);
+    return Application::provider('solital-password')->verify($value, $hash);
 }